requirements.txt
channels-redis
daphne

django~=5.0.4
django-viewflow~=2.0.2
asgiref~=3.8.1
channels~=4.1.0
pillow~=10.3.0
django-allauth~=0.62.1
python-dotenv~=1.0.1
djangorestframework~=3.15.1